requirements.txt
langchain
pypdf2
python-dotenv
streamlit
OpenAI
tiktoken
faiss-cpu